+++ spec/lib/verify/redacted_response_spec.rb
@@ -30,20 +30,20 @@ RSpec.describe Verify::RedactedResponse do
       "levelOfAssurance" => "LEVEL_2",
       "attributes" => {
         "firstNames" => [
-          {"value" => "XXXXXXXX", "verified" => true, "from" => "2019-06-25", "to" => "2019-06-30"},
-          {"value" => "XXXXXX", "verified" => true, "from" => "2015-03-01", "to" => "2019-06-25"},
-          {"value" => "XXX", "verified" => false, "from" => "2019-06-30"},
+          {"value" => "XXXXXXXX", "verified" => true, "from" => "XXXXXXXXXX", "to" => "2019-06-30"},
+          {"value" => "XXXXXX", "verified" => true, "from" => "XXXXXXXXXX", "to" => "2019-06-25"},
+          {"value" => "XXX", "verified" => false, "from" => "XXXXXXXXXX"},
         ],
         "middleNames" => [
-          {"value" => "XXXXXXX", "verified" => true, "from" => "2019-06-25", "to" => "2019-06-25"},
-          {"value" => "XX", "verified" => false, "from" => "2019-06-25"},
+          {"value" => "XXXXXXX", "verified" => true, "from" => "XXXXXXXXXX", "to" => "2019-06-25"},
+          {"value" => "XX", "verified" => false, "from" => "XXXXXXXXXX"},
         ],
         "surnames" => [
-          {"value" => "XXXXXXX", "verified" => true, "from" => "2015-03-01", "to" => "2019-06-24"},
-          {"value" => "XXXXXX", "verified" => true, "from" => "2019-06-25", "to" => "2019-06-25"},
-          {"value" => "", "verified" => false, "from" => "2019-06-25"},
+          {"value" => "XXXXXXX", "verified" => true, "from" => "XXXXXXXXXX", "to" => "2019-06-24"},
+          {"value" => "XXXXXX", "verified" => true, "from" => "XXXXXXXXXX", "to" => "2019-06-25"},
+          {"value" => "", "verified" => false, "from" => "XXXXXXXXXX"},
         ],
-        "datesOfBirth" => [{"value" => "XXXXXXXXXX", "verified" => true, "from" => "1806-04-09"}],
+        "datesOfBirth" => [{"value" => "XXXXXXXXXX", "verified" => true, "from" => "XXXXXXXXXX"}],
         "gender" => {"value" => "XXXX", "verified" => true},
         "addresses" => [
           {
@@ -52,7 +52,7 @@ RSpec.describe Verify::RedactedResponse do
               "postCode" => "XXXXXXX",
             },
             "verified" => true,
-            "from" => "2019-06-25",
+            "from" => "XXXXXXXXXX",
             "to" => "2019-06-25",
           },
           {
@@ -61,7 +61,7 @@ RSpec.describe Verify::RedactedResponse do
               "postCode" => "XXXXXXX",
             },
             "verified" => false,
-            "from" => "2019-06-25",
+            "from" => "XXXXXXXXXX",
           },
         ],
       },